esvdev_blog/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.11.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
cryptography==41.0.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.7
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-filter==23.3
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
Fabric3==1.14.post1
filetype==1.2.0
html5lib==1.1
idna==3.4
invoke==2.2.0
l18n==2021.3
openpyxl==3.1.2
paramiko==2.12.0
Pillow==10.1.0
pillow-heif==0.13.1
pycparser==2.21
PyNaCl==1.5.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.8.0
urllib3==2.1.0
wagtail==5.2
wagtail-cloudinary-image==0.0.3
webencodings==0.5.1
Willow==1.6.2
wrapt==1.16.0
